Create an abstract BridgeLocator for common behavior
